adds unique url identifier to the file name of each image
The last 14 characters of the url contain unique identifier (e.g. "EgF98uZt_zzGeg" from "https://artsandculture.google.com/asset/the-angel-of-the-annunciation/EgF98uZt_zzGeg"). This prevents accidental overwriting of files with the same title as often happens with duplicate titles like 'mountainous landscape' etc. that belong to the same artist by adding that unique string to the end of the file name.